Change Image containing content to BufferedImage (not tested yet)
git-svn-id: http://isdb1.offis.uni-oldenburg.de/repos/odysseus/trunk@21270 aa0bda67-78ed-0310-a382-c5338b0f89ba

Former-commit-id: 289fb6f5454cc33faa6c4fed71e76a7649f47aa3

+++ server/image/de.uniol.inf.is.odysseus.image/src/de/uniol/inf/is/odysseus/image/functions/FillImageFunction.java
@@ -47,7 +47,7 @@ public class FillImageFunction extends AbstractFunction<Image> {
 
 		Objects.requireNonNull(image);
 		
-        image.fill(value);
+        image.fill((int)value);
         return image;
     }
 }
